+++ src/Resources/SimulationRuns/Operations/Includes.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paddle\SDK\Resources\SimulationRuns\Operations;
+
+use Paddle\SDK\PaddleEnum;
+
+/**
+ * @method static Includes Events()
+ */
+class Includes extends PaddleEnum
+{
+    public const Events = 'events';
+}